+++ CHANGELOG.md
@@ -38,6 +38,7 @@ title: Changelog
 * `<cc-select>`: 
   * remove the unique id generation technique and rely on Shadow DOM isolation instead.
   * only disable the placeholder option if the component is in required mode.
+  * add red border and redish focus ring when error slot is used.
 * `<cc-toggle>`: remove the unique name generation technique and rely on Shadow DOM isolation instead.
 * `accessibility Styles`: add new `accessibilityStyles` containing a `visually-hidden` class to hide content visually but not from assistive technologies.
 * `<cc-env-var-create>`: add visually hidden label for all input fields so that these fields can be identified by assistive technologies.